version filters for node args
